Add TypedRequest* types
These can be used for cases where the actual request handler and the
middleware provided by these packages aren't in the same file: in this
case, Typescript doesn't automatically infer the types and display them
in your editor for you. You can pass your ZodType to the TypedRequest
type, and it will give you a typed request for you.